requirements.txt
certifi==2021.5.30
charset-normalizer==2.0.4
click==8.0.1
colorama==0.4.4
Flask==2.0.1
Flask-Cors==3.0.10
idna==3.2
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
numpy==1.21.2
opencv-python==4.5.3.56
Pillow==8.3.1
pytesseract==0.3.8
requests==2.26.0
six==1.16.0
urllib3==1.26.6
Werkzeug==2.0.1
